progressed embedder

+++ data_utils/embedder.py
@@ -73,15 +73,38 @@ def _load_cached_documents(cache_dir: str) -> list[Document]:
             docs.append(doc)
     return docs
 
-def _embedding_strategy_raw(docs: list[Document], embeddings_model: Embeddings, result_path: str):
+def _record_current_embedded_doc(doc_index: int, result_path: str):
+
+    with open(f"{result_path}/embedding_metadata.json", "r") as f:
+        metadata = json.load(f)
+    metadata["current_embedded_doc"] = doc_index
+    with open(f"{result_path}/embedding_metadata.json", "w") as f:
+        json.dump(metadata, f)
+
+def _load_current_embedded_doc_index(result_path: str) -> int:
+    with open(f"{result_path}/embedding_metadata.json") as f:
+        metadata = json.load(f)
+        return metadata.get("current_embedded_doc", 0)
+
+def _batch_add_documents(vectorstore: Chroma, docs: list[Document], result_path: str, batch_size: int=16):
+    start_index = _load_current_embedded_doc_index(result_path)
+    progress = tqdm(total=len(docs)-start_index, desc="Adding documents to vectorstore")
+    for i in range(start_index, len(docs), batch_size):
+        vectorstore.add_documents(docs[i:i+batch_size])
+        _record_current_embedded_doc(i+batch_size, result_path)
+        progress.update(batch_size)
+
+
+def _embedding_strategy_raw(docs: list[Document], embeddings_model: Embeddings, result_path: str, batch_size: int=16):
     from langchain_text_splitters import RecursiveCharacterTextSplitter
     text_splitter = RecursiveCharacterTextSplitter(
             chunk_size=500, chunk_overlap=100, add_start_index=True
         )
     all_splits = text_splitter.split_documents(docs)
-    Chroma.from_documents(documents=all_splits, embedding=embeddings_model, persist_directory=result_path)
+    vectorstore = Chroma(persist_directory=result_path, embedding_function=embeddings_model)
+    _batch_add_documents(vectorstore, all_splits, result_path, batch_size)
     
-def _embedding_strategy_hypothetical_question(docs: list[Document], embeddings_model: Embeddings, result_path: str,chat_model_name: str="Qwen/Qwen2.5-1.5B-Instruct"):
+def _embedding_strategy_hypothetical_question(docs: list[Document], embeddings_model: Embeddings, result_path: str, batch_size: int=16,chat_model_name: str="Qwen/Qwen2.5-1.5B-Instruct"):
     def generate_hypothetical_question(doc: Document, llm):
         from langchain_core.prompts import PromptTemplate
         template = PromptTemplate(
@@ -118,9 +141,10 @@ def _embedding_strategy_hypothetical_question(docs: list[Document], embeddings_m
     else:
         print("Hypothetical questions loaded from cache")
     print("Embedding hypothetical questions")
-    Chroma.from_documents(documents=hypothetical_questions, embedding=embeddings_model, persist_directory=result_path)
+    vectorstore = Chroma(persist_directory=result_path, embedding_function=embeddings_model)
+    _batch_add_documents(vectorstore, hypothetical_questions, result_path, batch_size)
 
-def _embedding_strategy_hypothetical_question_with_raw(docs: list[Document], embeddings_model: Embeddings, result_path: str,chat_model_name: str="Qwen/Qwen2.5-1.5B-Instruct"):
+def _embedding_strategy_hypothetical_question_with_raw(docs: list[Document], embeddings_model: Embeddings, result_path: str, batch_size: int=16,chat_model_name: str="Qwen/Qwen2.5-1.5B-Instruct"):
     def generate_hypothetical_question(doc: Document, llm):
         from langchain_core.prompts import PromptTemplate
         template = PromptTemplate(
@@ -158,17 +182,20 @@ def _embedding_strategy_hypothetical_question_with_raw(docs: list[Document], emb
     else:
         print("Hypothetical question docs loaded from cache")
     print("Embedding hypothetical question docs")
-    Chroma.from_documents(documents=hypothetical_questions, embedding=embeddings_model, persist_directory=result_path)
+    vectorstore = Chroma(persist_directory=result_path, embedding_function=embeddings_model)
+    _batch_add_documents(vectorstore, hypothetical_questions, result_path, batch_size)
 
 EmbeddingStrategy = Literal["hypothetical_question", "raw", "hypothetical_question_with_raw"]
-def save_vectorstore_from_huggingface(content_json_path: str, result_path: str, embedding_model_name: str, embedding_strategy: EmbeddingStrategy="hypothetical_question"):
+def save_vectorstore_from_huggingface(content_json_path: str, result_path: str, embedding_model_name: str, embedding_strategy: EmbeddingStrategy="hypothetical_question", batch_size: int=16):
     """
     Create vectorstore from content_json_path (created from extract_content) with embedding_model_name; save the results result_path
     """
     if not os.path.exists(result_path):
         os.makedirs(result_path)
-    with open(f"{result_path}/embedding_metadata.json", "w") as f:
-        json.dump({"embedding_model_name": embedding_model_name}, f)
+    embedding_metadata_json = f"{result_path}/embedding_metadata.json"
+    if not os.path.exists(embedding_metadata_json):
+        with open(embedding_metadata_json, "w") as f:
+            json.dump({"embedding_model_name": embedding_model_name}, f)
     loader = HTMLJSONLoader(content_json_path)
     docs = loader.load()
     embeddings_model = HuggingFaceEmbeddings(model_name=embedding_model_name, model_kwargs={"trust_remote_code": True})
@@ -181,4 +208,4 @@ def save_vectorstore_from_huggingface(content_json_path: str, result_path: str, 
         embedding_func = _embedding_strategy_hypothetical_question_with_raw
     else:
         raise ValueError(f"Unknown embedding strategy: {embedding_strategy}")
-    embedding_func(docs, embeddings_model, result_path)
+    embedding_func(docs, embeddings_model, result_path, batch_size)
